+++ src/main/java/it/unicam/cs/pa/jlogo/app/ObservableLogoCursor.java
@@ -3,9 +3,9 @@ package it.unicam.cs.pa.jlogo.app;
 import it.unicam.cs.pa.jlogo.LogoCursor;
 import it.unicam.cs.pa.jlogo.Point;
 import it.unicam.cs.pa.jlogo.model.Line;
-import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
 import java.awt.Color;
@@ -14,7 +14,7 @@ import java.util.Optional;
 public class ObservableLogoCursor extends LogoCursor {
 
     private final ObjectProperty<Point> position = new SimpleObjectProperty<>(this, "position");
-    private final IntegerProperty direction = new SimpleIntegerProperty(this, "direction");
+    private final DoubleProperty direction = new SimpleDoubleProperty(this, "direction");
     private final ObjectProperty<Color> lineColor = new SimpleObjectProperty<>(this, "lineColor");
 
 
@@ -28,19 +28,19 @@ public class ObservableLogoCursor extends LogoCursor {
 
 
     @Override
-    public Optional<Line> move(int distance) {
+    public Optional<Line> move(double distance) {
         Optional<Line> result = super.move(distance);
         position.set(getPosition());
         return result;
     }
 
     @Override
-    public void rotate(int degrees) {
+    public void rotate(double degrees) {
         direction.set((getDirection() + degrees) % 360);
     }
 
     @Override
-    public int getDirection() {
+    public double getDirection() {
         return direction.get();
     }
 
@@ -54,7 +54,7 @@ public class ObservableLogoCursor extends LogoCursor {
         return position;
     }
 
-    public IntegerProperty directionProperty() {
+    public DoubleProperty directionProperty() {
         return direction;
     }
